[Refactor] Establish data sink module boundaries
