+++ be/CMakeLists.txt
@@ -1013,6 +1013,7 @@ add_subdirectory(${SRC_DIR}/exec_primitive)
 add_subdirectory(${SRC_DIR}/connector_primitive)
 add_subdirectory(${SRC_DIR}/connector)
 add_subdirectory(${SRC_DIR}/module)
+add_subdirectory(${SRC_DIR}/data_sink)
 add_subdirectory(${SRC_DIR}/exec)
 add_subdirectory(${SRC_DIR}/schema_scanner)
 add_subdirectory(${SRC_DIR}/exprs)
