+++ be/module_boundary_manifest.json
@@ -372,6 +372,33 @@
       "remediation": "Keep format code inside the Format module; move connector orchestration and higher execution policy upward."
     },
     {
+      "id": "datasinkfile",
+      "doc_label": "DataSinkFile",
+      "summary": "Reusable file sink builders for delimited text and Parquet output over explicit format, filesystem, expression, and column dependencies without full Exec, concrete Storage, or orchestration coupling.",
+      "agents_path": "be/src/data_sink/AGENTS.md",
+      "owned_targets": ["DataSinkFile"],
+      "owned_roots": ["be/src/data_sink/file"],
+      "allowed_include_prefixes": [
+        "data_sink/file/",
+        "formats/",
+        "fs/",
+        "exprs/",
+        "runtime/",
+        "column/",
+        "types/",
+        "common/",
+        "base/",
+        "gutil/",
+        "gen_cpp/"
+      ],
+      "forbidden_include_prefixes": ["exec/", "storage/", "service/", "connector/", "agent/", "http/", "compute_env/", "cache/", "data_workflows/", "orchestration/"],
+      "forbidden_includes": ["exec/exec_env.h"],
+      "allowed_target_deps": ["FormatCore", "FormatCsv", "FormatParquet", "Expr", "FileSystem", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
+      "allowed_test_targets": ["data_sink_file_test"],
+      "allowed_test_link_deps": ["DataSinkFile", "FormatCore", "FormatCsv", "FormatParquet", "Expr", "Runtime", "FileSystem", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
+      "remediation": "Keep file serialization builders independent of full Exec and concrete Storage; place sink lifecycle and orchestration in higher data-sink modules instead of expanding this leaf target."
+    },
+    {
       "id": "storageprimitive",
       "doc_label": "StoragePrimitive",
       "summary": "Shared storage-access primitives, predicate contracts, predicate trees, iterator/merge/aggregation helpers, and value/schema types used by Storage, Exec, Connector, Format, and ComputeEnv without concrete Storage engine, tablet, rowset, lake, service, or full Exec coupling.",
@@ -702,7 +729,7 @@
     {
       "id": "execprimitive",
       "doc_label": "ExecPrimitive",
-      "summary": "Primitive execution contracts, DataSink base contract, runtime-filter infrastructure, generic morsel queues, and stable pipeline operator/factory primitives without broader Exec runtime, scheduler, storage, service, or connector coupling.",
+      "summary": "Primitive execution contracts, synchronous and asynchronous DataSink lifecycle contracts, runtime-filter infrastructure, generic morsel queues, and stable pipeline operator/factory primitives without broader Exec runtime, scheduler, storage, service, or connector coupling.",
       "owned_targets": ["ExecPrimitive"],
       "owned_roots": ["be/src/exec_primitive"],
       "allowed_include_prefixes": [
@@ -729,7 +756,7 @@
       "allowed_target_deps": ["Expr", "Runtime", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
       "allowed_test_targets": ["exec_primitive_test"],
       "allowed_test_link_deps": ["ExecPrimitive", "Expr", "Runtime", "ChunkCore", "ColumnCore", "Types", "Common", "Base", "Gutil", "StarRocksGen"],
-      "remediation": "Keep ExecPrimitive limited to execution contracts, the DataSink base contract, runtime filters, generic morsel queues, and stable operator/factory primitives; move runtime, scheduler, concrete operators, concrete sinks, storage, service, and connector integration upward."
+      "remediation": "Keep ExecPrimitive limited to execution contracts, synchronous and asynchronous DataSink lifecycle contracts, runtime filters, generic morsel queues, and stable operator/factory primitives; move runtime, scheduler, concrete operators, concrete sinks, storage, service, and connector integration upward."
     },
     {
       "id": "execruntime",
