+++ be/src/exec/data_sinks/file_result_writer.cpp
@@ -40,9 +40,8 @@
 #include "base/uid_util.h"
 #include "column/chunk.h"
 #include "common/config_rpc_client_fwd.h"
-#include "exec/local_file_writer.h"
-#include "exec/parquet_builder.h"
-#include "exec/plain_text_builder.h"
+#include "data_sink/file/parquet_builder.h"
+#include "data_sink/file/plain_text_builder.h"
 #include "formats/csv/converter.h"
 #include "formats/io/formatted_output_stream.h"
 #include "fs/fs_factory.h"
