+++ be/src/exec/data_sinks/file_result_writer.h
@@ -35,10 +35,9 @@
 #pragma once
 
 #include "exec/data_sinks/result_writer.h"
-#include "exec/parquet_builder.h"
+#include "formats/parquet/file_writer.h"
 #include "fs/fs_fwd.h"
 #include "gen_cpp/DataSinks_types.h"
-#include "parquet/file_writer.h"
 
 namespace starrocks {
