+++ be/src/exec/data_sinks/tablet_sink.h
@@ -35,6 +35,7 @@
 #pragma once
 
 #include "exec/data_sinks/tablet_sink_sender.h"
+#include "exec_primitive/async_data_sink.h"
 
 namespace starrocks {
