+++ be/src/exec/data_sinks/tablet_sink_index_channel.h
@@ -32,7 +32,6 @@
 #include "common/status.h"
 #include "common/thread/threadpool.h"
 #include "common/tracer_fwd.h"
-#include "exec/data_sinks/async_data_sink.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "runtime/mem_tracker.h"
