+++ be/src/exec/data_sinks/tablet_sink_sender.cpp
@@ -20,10 +20,10 @@
 #include "column/chunk.h"
 #include "common/config_ingest_fwd.h"
 #include "common/statusor.h"
-#include "exec/write_combined_txn_log.h"
 #include "exprs/expr.h"
 #include "exprs/expr_executor.h"
 #include "runtime/runtime_state.h"
+#include "storage/lake/combined_txn_log_writer.h"
 
 namespace starrocks {
