+++ be/src/exec/pipeline/sink/export_sink_operator.cpp
@@ -17,11 +17,11 @@
 #include "base/uid_util.h"
 #include "compute_env/workgroup/pipeline_executor_set.h"
 #include "compute_env/workgroup/work_group.h"
-#include "exec/file_builder.h"
+#include "data_sink/file/file_builder.h"
+#include "data_sink/file/plain_text_builder.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/fragment_context_cancel.h"
 #include "exec/pipeline/sink/sink_io_buffer.h"
-#include "exec/plain_text_builder.h"
 #include "exec_primitive/data_sink.h"
 #include "exec_primitive/pipeline/primitives/driver_executor.h"
 #include "exprs/expr_executor.h"
